refactor(types): move all shared types into `@types`
