+++ src/combinators.ts
@@ -1,6 +1,5 @@
 export * from '@combinators/chain'
 export * from '@combinators/choice'
-export * from '@combinators/when'
 export * from '@combinators/error'
 export * from '@combinators/many'
 export * from '@combinators/map'
@@ -9,3 +8,4 @@ export * from '@combinators/sepBy'
 export * from '@combinators/sequence'
 export * from '@combinators/take'
 export * from '@combinators/until'
+export * from '@combinators/when'
