+++ src/combinators/chain.ts
@@ -2,7 +2,7 @@ import { many } from './many'
 import { map } from './map'
 import { sequence } from './sequence'
 
-import type { Parser } from '@state'
+import type { Parser } from '@types'
 
 /** @internal */
 type Fn<L, R> = (left: L, right: R) => L
