+++ src/combinators/choice.ts
@@ -1,5 +1,4 @@
-import type { Parser } from '@state'
-import type { ToUnion } from '@types'
+import type { Parser, ToUnion } from '@types'
 
 /**
  * Applies `ps` parsers in order until one of them succeeds.
