+++ src/combinators/error.ts
@@ -1,4 +1,4 @@
-import type { Parser } from '@state'
+import type { Parser } from '@types'
 
 /**
  * Replaces `parser`'s error message with `expected`.
