+++ src/index.ts
@@ -1,3 +1,3 @@
 export * from '@combinators'
 export * from '@parsers'
-export * from '@state'
+export * from '@types'
