+++ src/parsers/eol.ts
@@ -1,8 +1,7 @@
 import { string } from './string'
 
-import { choice } from '@combinators/choice'
-import { error } from '@combinators/error'
-import type { Parser } from '@state'
+import { choice, error } from '@combinators'
+import type { Parser } from '@types'
 
 const EOL_UNIX = '\n'
 const EOL_NON_UNIX = '\r\n'
