+++ src/parsers/whitespace.ts
@@ -1,6 +1,6 @@
 import { regexp } from './regexp'
 
-import type { Parser } from '@state'
+import type { Parser } from '@types'
 
 const WHITESPACE_REQUIRED_RE = /\s+/g
